% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cpi.R
\name{cpi}
\alias{cpi}
\title{AEDForecasting:
Package to incorporate change point analysis in ARIMA forecasting}
\usage{
cpi(myts, startChangePoint = 1, endChangePoint = 0, step = 1, num = 8,
  cpmeth = "BinSeg", CPpenalty = "SIC")
}
\arguments{
\item{myts}{a time series object}

\item{startChangePoint}{a positive integer for minimum number of changepoints}

\item{endChangePoint}{a positive integer for maximum number of change points. If 0 then  only startChangePoint number of change points will be entered. Should be either 0 or greater than startChangePoint and if so the algorithm will loop through all values inbetween subject to step}

\item{step}{an integer to step through loop of change points}

\item{num}{Bump model number (see below)}

\item{cpmeth}{changepoint method. Default is BinSeg. See cpa package for details}

\item{CPpenalty}{default is SIC. See cpa package for details}
}
\value{
The model number, number of change points, and the information criteria values

The bump models are described below.
}
\description{
AEDForecasting:
Package to incorporate change point analysis in ARIMA forecasting
}

